requirements.txt
flask==3.0.3
groq>=1.0.0
python-dotenv==1.0.1
gunicorn==21.2.0
